requirements.txt
airtable-python-wrapper==0.11.3.post1
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
Django==2.1.3
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.1.1
pytz==2018.7
requests==2.21.0
six==1.11.0
urllib3==1.24.1
wincertstore==0.2
wrapt==1.10.11
